testing modularized functions
